kafka/requirements.txt
awscli==1.36.7
boto3==1.35.66
botocore==1.35.66
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
confluent-kafka==2.6.1
docutils==0.16
idna==3.10
jmespath==1.0.1
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.1.0
pyasn1==0.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
